0.4.0412
1.实现地图无线连接（大改scene关于map的地方）
2.金币碰撞及计数实现
3.修复毛虫以及绳子碰撞区域

未实现：
1.毛虫属于下一个map
2.好多好多。。。。

bug:
1.一个金币 计数两次。
2.毛虫在当前绳子没有更新，导致一直向上爬。

+++ Controller/FallDownController.h
@@ -21,6 +21,6 @@ public:
 	/* 设置Y方向的移动速度*/
 	void setiYSpeed(int iSpeed);
 private:
-	int iYSpeed;
+	float iYSpeed;
 };
